requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
black==19.10b0
colorama==0.4.4
coverage==5.3
DateTime==4.3
flake8==3.8.4
importlib-metadata==3.1.0
jsonpickle==1.4.1
more-itertools==8.6.0
MouseInfo==0.1.3
numpy==1.19.4
packaging==20.4
Pillow==9.0.0
pluggy==0.13.1
py==1.9.0
PyAutoGUI==0.9.52
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.1
pyparsing==2.4.7
pyperclip==1.8.1
PyRect==0.1.4
PyScreeze==0.1.26
pyserial==3.5
pystray==0.17.1
pytest==6.1.2
pytest-cov==2.10.1
python-xlib==0.29
python3-xlib==0.15
PyTweening==1.0.3
pytz==2020.4
six==1.15.0
svgwrite==1.4
wasabi==0.8.0
wcwidth==0.2.5
zipp==3.4.0
zope.interface==5.2.0
streamlit==0.57.3
python-osc==1.7.4
mlrose-hiive==2.1.3
fpdf2==2.3.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyqt5==5.15.4
scipy==1.6.0